Fix for hud disappear
After a few tests, it turned out that the position of the HUD is completely different from the cursor position and it is impossible to save the data with the coordinates 1x1. So I set the maximum ranges and additionally added an error for the Y position. Checked on FHD, 2K and 4K resolutions and only at 4K is the Y position error visible (after loading `+0.02361` [51px]), but after adjusting for this error on the FHD, the HUD is off the top edge.

4K, set and saved:
https://cdn.discordapp.com/attachments/939967467931975782/987483978296623174/2022-06-17_18_27_34-Farming_Simulator_22.png
loaded:
https://cdn.discordapp.com/attachments/939967467931975782/987483979592630312/2022-06-17_18_28_29-Farming_Simulator_22.png